cityos_json/requirements.txt
elasticsearch
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
starlette
aiofiles
mysqlclient==2.1.1
requests
datetime
chardet
pydantic-settings
python-dotenv
boto3